FEATURE: Add trigger_regen pipe to run loop for manual regenerate_config calls (#1)
This adds a @trigger_regen_w and @trigger_regen_r IO pipe generated at initialize of the config class. Consumers can write to this pipe to trigger a regeneration of the config file and a force reload of the server. This is an alternative to having to watch a file for writes.

The config class exposes a regen_notifier method which returns an instance of ConfigRegenNotifier which holds the instance variable holding the write end of the pipe. When a consumer wants to trigger a config regeneration, all they need to do is this:

```
notif = config.regen_notifier
trigger.trigger_regen
```

Also add specs for the other IO.select objects in the run loop (file watches and shutdown pipe).