FEATURE: Add cooldown_duration to reduce regen frequency
Hammering a downstream service with reload requests is often a bad idea.
This method exists to allow subclasses to define a 'cooldown' duration.
After each config regeneration, the config generator will sleep for this
duration, regardless of any CONT signals or inotify events. Those events
will be queued up, and processed at the end of the cooldown.

Defaults to 5 seconds, but can be altered by sub classes